have a dialect selection box in telemetry selection
