+++ libs/src/stores/dialect.ts
@@ -1,4 +1,5 @@
-import { createWithEqualityFn as create } from 'zustand/traditional'
+import { create } from 'zustand'
+import { createJSONStorage, persist } from 'zustand/middleware'
 import { Dialect } from '@libs/dialects/dialect'
 import { DialectCommandDescription } from '@libs/commands/command'
 import { dialectRegistry, DEFAULT_DIALECT_ID } from '@libs/dialects/dialectRegistry'
@@ -20,16 +21,30 @@ type Actions = {
 const defaultDialect =
   dialectRegistry.find((d) => d.id === DEFAULT_DIALECT_ID) ?? dialectRegistry[0]
 
-export const useDialect = create<State & Actions>((set) => ({
-  activeDialect: defaultDialect,
-  activeDialectId: defaultDialect.id,
+export const useDialect = create<State & Actions>()(
+  persist(
+    (set) => ({
+      activeDialect: defaultDialect,
+      activeDialectId: defaultDialect.id,
 
-  setDialect: (id) => {
-    const dialect = dialectRegistry.find((d) => d.id === id)
-    if (!dialect) {
-      console.warn(`[useDialect] Unknown dialect id: "${id}"`)
-      return
+      setDialect: (id) => {
+        const dialect = dialectRegistry.find((d) => d.id === id)
+        if (!dialect) {
+          console.warn(`[useDialect] Unknown dialect id: "${id}"`)
+          return
+        }
+        set({ activeDialect: dialect, activeDialectId: id })
+      }
+    }),
+    {
+      name: 'dialect-storage',
+      storage: createJSONStorage(() => localStorage),
+      partialize: (state) => ({ activeDialectId: state.activeDialectId }),
+      onRehydrateStorage: () => (state) => {
+        if (!state) return
+        const dialect = dialectRegistry.find((d) => d.id === state.activeDialectId)
+        if (dialect) state.activeDialect = dialect
+      }
     }
-    set({ activeDialect: dialect, activeDialectId: id })
-  }
-}))
+  )
+)
